VendorManagementSystemBackEnd/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
mssql-django==1.5
mysqlclient==2.2.4
pip==23.2.1
pyodbc==5.1.0
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
